virtual-workspaces: move client+informer setup to command
